add go build flag to skip for window
